Disable language prefix in URLs
We don't have multiple languages, yes, and this URL system makes things complicated. E.g. with redirects, that would have to be dynamic.
