+++ components/Section/ContactSection.tsx
@@ -44,7 +44,7 @@ export default function ContactSection() {
               </h4>
               <p>{t('homepage.getInTouch.events.description')}</p>
               <p>
-                <Link href={`/[lang]#events`} as={`/${locale}#events`}>
+                <Link href={`/#events`}>
                   <a className='highlight'>
                     {t('homepage.getInTouch.events.cta')}
                   </a>
