+++ pages/about.tsx
@@ -1,11 +1,11 @@
 import { NextPage } from 'next'
-import PageLayout from '../../components/PageLayout/PageLayout'
+import PageLayout from '../components/PageLayout/PageLayout'
 import Grid from '@material-ui/core/Grid'
-import TextSection from '../../components/Section/TextSection'
-// import LocalSwitcher from '../../components/LocalSwitcher/LocalSwitcher'
-import Button from '../../components/Button/Button'
-import useTranslation from '../../hooks/useTranslation'
-import WithLocale from '../../containers/withLocale'
+import TextSection from '../components/Section/TextSection'
+// import LocalSwitcher from '../components/LocalSwitcher/LocalSwitcher'
+import Button from '../components/Button/Button'
+import useTranslation from '../hooks/useTranslation'
+import WithLocale from '../containers/withLocale'
 
 const About: NextPage = () => {
   const { t } = useTranslation()
